Rename `report_functions` to `report`.
