+++ rust/src/nasl/builtin/report/mod.rs
