+++ rust/src/nasl/builtin/report/tests.rs
